Stashing: If an error occured after stashing, remember to pop the stashed changes.
Additionally, it made sense to move this into a lib.rs function.
